Renames `AssetSequenceComponentLoader*` to `*Ascl`.
