+++ crate/loading/src/lib.rs
@@ -9,13 +9,10 @@ pub use crate::{
         AssetDefinitionLoader, AssetDefinitionLoadingSystem, AssetDiscoverySystem,
         AssetDiscoverySystemData, AssetIdMapper, AssetIdMappingSystem, AssetPartLoader,
         AssetPartLoadingCoordinatorSystem, AssetPartLoadingCoordinatorSystemData,
-        AssetPartLoadingSystem, AssetSequenceComponentLoader,
-        AssetSequenceComponentLoaderUiCharacterSelection, AssetSequenceComponentLoaderUiComponents,
-        AssetSequenceComponentLoaderUiControlSettings, AssetSequenceComponentLoaderUiForm,
-        AssetSequenceComponentLoaderUiMapSelection, AssetSequenceComponentLoaderUiMenu,
-        AssetSequenceComponentLoaderUiSessionLobby, AssetSequenceComponentLoadingSystem,
+        AssetPartLoadingSystem, AssetSequenceComponentLoader, AssetSequenceComponentLoadingSystem,
         AssetSpritesDefinitionLoader, AssetSpritesDefinitionLoadingSystem, AssetTextureLoader,
-        AssetTextureLoadingSystem,
+        AssetTextureLoadingSystem, UiCharacterSelectionAscl, UiComponentsAscl,
+        UiControlSettingsAscl, UiFormAscl, UiMapSelectionAscl, UiMenuAscl, UiSessionLobbyAscl,
     },
     system_data::{
         AssetLoadingResources, DefinitionLoadingResources, DefinitionLoadingResourcesRead,
