+++ crate/loading/src/system/asset_sequence_component_loading_system/map_ascl.rs
@@ -18,9 +18,9 @@ use crate::{
 
 /// Loads sequence components for map assets.
 #[derive(Debug)]
-pub struct AssetSequenceComponentLoaderMap;
+pub struct MapAscl;
 
-impl AssetSequenceComponentLoaderMap {
+impl MapAscl {
     /// Loads sequence components for map assets.
     pub fn load(
         asset_loading_resources: &mut AssetLoadingResources<'_>,
