+++ crate/loading/src/system/asset_sequence_component_loading_system/ui_components_ascl.rs
@@ -6,7 +6,7 @@ use sprite_model::loaded::{
 
 /// Common components for UI items.
 #[derive(Clone, Debug)]
-pub struct AssetSequenceComponentLoaderUiComponents {
+pub struct UiComponentsAscl {
     /// Sequence transition upon sequence end.
     pub sequence_end_transitions: SequenceEndTransitions,
     /// Sequence of `WaitSequenceHandle`s.
